refactor: remove existing assets querying
