+++ test/writeClient-migrate-assets.test.ts
@@ -17,83 +17,6 @@ const isNode16 = process.version.startsWith("v16")
 const isNode18 = process.version.startsWith("v18")
 const it = _it.skipIf(isNode16 || isNode18)
 
-it.concurrent("discovers existing assets", async (ctx) => {
-	const client = createTestWriteClient({ ctx })
-
-	mockPrismicRestAPIV2({ ctx })
-	mockPrismicAssetAPI({ ctx, client, existingAssets: [2] })
-	mockPrismicMigrationAPI({ ctx, client })
-
-	const migration = prismic.createMigration()
-
-	const reporter = vi.fn()
-
-	await client.migrate(migration, { reporter })
-
-	expect(reporter).toHaveBeenCalledWith({
-		type: "assets:existing",
-		data: {
-			existing: 2,
-		},
-	})
-})
-
-it.concurrent(
-	"discovers existing assets and crawl pages if any",
-	async (ctx) => {
-		const client = createTestWriteClient({ ctx })
-
-		mockPrismicRestAPIV2({ ctx })
-		mockPrismicAssetAPI({ ctx, client, existingAssets: [2, 3] })
-		mockPrismicMigrationAPI({ ctx, client })
-
-		const migration = prismic.createMigration()
-
-		const reporter = vi.fn()
-
-		await client.migrate(migration, { reporter })
-
-		expect(reporter).toHaveBeenCalledWith({
-			type: "assets:existing",
-			data: {
-				existing: 5,
-			},
-		})
-	},
-)
-
-it.concurrent("skips creating existing assets", async (ctx) => {
-	const client = createTestWriteClient({ ctx })
-
-	mockPrismicRestAPIV2({ ctx })
-	const { assetsDatabase } = mockPrismicAssetAPI({
-		ctx,
-		client,
-		existingAssets: [1],
-	})
-	mockPrismicMigrationAPI({ ctx, client })
-
-	const migration = prismic.createMigration()
-
-	const asset = assetsDatabase[0][0]
-	const migrationAsset = migration.createAsset(asset)
-
-	const reporter = vi.fn()
-
-	await client.migrate(migration, { reporter })
-
-	expect(reporter).toHaveBeenCalledWith({
-		type: "assets:skipping",
-		data: {
-			reason: "already exists",
-			current: 1,
-			remaining: 0,
-			total: 1,
-			asset: migrationAsset,
-		},
-	})
-})
-
 it.concurrent("creates new asset from string file data", async (ctx) => {
 	const client = createTestWriteClient({ ctx })
